src/AuthSession/ICipher.py
from Crypto.Cipher import AES
from abc import ABC, abstractmethod


class ICipher(ABC):
    def __init__(self, key):
        self.key = key
        self.cipher = AES.new(self.key, AES.MODE_ECB)

    @abstractmethod
    def encrypt(self, data):
        pass

    @abstractmethod
    def decrypt(self, data):
        pass
