requirements.txt
requests~=2.28.1
pycryptodome~=3.15.0